moved storage into helpers and fixed all references
